latseq: upgrade data types to 64-bit and double ring buffer size
- Upgrade data identifiers in `latseq_element_t` and logging macros
  from 32-bit (int32_t) to 64-bit (int64_t) to handle larger ID ranges.
- Increase `RING_BUFFER_SIZE` from 1024 to 2048 to prevent overflows
  under high load, maintaining a power-of-two size to preserve fast
  bitwise modulo operations (`& (SIZE - 1)`).
- Update format specifiers in all `LATSEQ_P` macro invocations across
  the codebase (converting `%d` to `%ld`, etc.) to prevent `sprintf`
  stack alignment corruption with the new 64-bit types.

+++ common/utils/LATSEQ/latseq.c
@@ -205,7 +205,7 @@ static int write_latseq_entry(void)
   free(tmps);
   // cleanup buffer element
   e->ts = 0;
-  memset(e->data_id, 0, (sizeof(uint32_t) * e->len_id));
+  memset(e->data_id, 0, (sizeof(int64_t) * e->len_id));
   e->len_id = 0;
   
   //Update read_head for the current read_ith_thread
